+++ R/dm.R
@@ -763,7 +763,7 @@ glimpse.dm <- function(x, width = NULL, ...) {
   }
   cat_line(
     trim_width(
-      paste0("dm of ", length(table_list), " tables: ", paste0(tick(names(table_list)), collapse = ", ")),
+      paste0("dm of ", length(table_list), " tables: ", toString(tick(names(table_list)))),
       glimpse_width
     )
   )
@@ -774,10 +774,11 @@ glimpse.dm <- function(x, width = NULL, ...) {
     pk <- dm_get_pk_impl(x, table_name) %>%
       map_chr(~ paste0("(", paste0(tick(.x), collapse = ", "), ")"))
     if (!is_empty(pk)) {
-      cat_line(trim_width(paste0(length(pk), " primary key(s): ", pk), glimpse_width))
+      # FIXME: needs to change if #622 is solved
+      cat_line(trim_width(paste0("Primary key: ", pk), glimpse_width))
     }
     fk <- all_fks %>%
-      filter(child_table == !!table_name) %>%
+      filter(child_table == table_name) %>%
       select(-child_table) %>%
       pmap_chr(
         function(child_fk_cols, parent_table, parent_key_cols) {