+++ R/financial.R
@@ -15,7 +15,10 @@
 #' dm_financial() %>%
 #'   dm_draw()
 dm_financial <- function() {
-  stopifnot(rlang::is_installed("RMariaDB"))
+  check_suggested("RMariaDB",
+    use = TRUE,
+    top_level_fun = "dm_financial"
+  )
 
   my_db <- financial_db_con()
 
@@ -67,7 +70,10 @@ dm_has_financial <- function() {
 #' @rdname dm_financial
 #' @export
 dm_financial_sqlite <- function() {
-  stopifnot(rlang::is_installed("RSQLite"))
+  check_suggested("RSQLite",
+    use = TRUE,
+    top_level_fun = "dm_financial_sqlite"
+  )
 
   my_dm <-
     dm_financial() %>%