+++ R/rows-db.R
@@ -30,6 +30,11 @@
 #'   Note that also columns not in `y` but automatically created when inserting
 #'   into `x` can be returned, for example the `id` column.
 #'
+#'   Due to upstream limitations, a warning is given if this argument
+#'   is passed unquoted.
+#'   To avoid the warning, quote the argument manually:
+#'   use e.g. `returning = quote(everything())` .
+#'
 #' @return A tbl object of the same structure as `x`.
 #'   If `in_place = TRUE`, the underlying data is updated as a side effect,
 #'   and `x` is returned, invisibly. If return columns are specified with
@@ -44,10 +49,13 @@
 #' try(rows_insert(data, tibble::tibble(a = 4, b = "z")))
 #' rows_insert(data, tibble::tibble(a = 4, b = "z"), copy = TRUE)
 #' rows_update(data, tibble::tibble(a = 2:3, b = "w"), copy = TRUE, in_place = FALSE)
+#' rows_patch(data, dbplyr::memdb_frame(a = 1:4, c = 0), in_place = FALSE)
 #'
 #' rows_insert(data, dbplyr::memdb_frame(a = 4, b = "z"), in_place = TRUE)
 #' data
 #' rows_update(data, dbplyr::memdb_frame(a = 2:3, b = "w"), in_place = TRUE)
+#' data
+#' rows_patch(data, dbplyr::memdb_frame(a = 1:4, c = 0), in_place = TRUE)
 #' data
 NULL
 
@@ -56,7 +64,15 @@ NULL
 rows_insert.tbl_dbi <- function(x, y, by = NULL, ...,
                                 in_place = NULL, copy = FALSE, check = NULL,
                                 returning = NULL) {
-  returning_cols <- eval_select_both(enquo(returning), colnames(x))$names
+
+  # Expect manual quote from user, silently fall back to enexpr()
+  returning_expr <- enexpr(returning)
+  tryCatch(
+    returning_expr <- returning,
+    error = identity
+  )
+
+  returning_cols <- eval_select_both(returning_expr, colnames(x))$names
   check_returning_cols_possible(returning_cols, in_place)
 
   y <- auto_copy(x, y, copy = copy)
@@ -91,7 +107,15 @@ rows_insert.tbl_dbi <- function(x, y, by = NULL, ...,
 rows_update.tbl_dbi <- function(x, y, by = NULL, ...,
                                 in_place = NULL, copy = FALSE, check = NULL,
                                 returning = NULL) {
-  returning_cols <- eval_select_both(enquo(returning), colnames(x))$names
+
+  # Expect manual quote from user, silently fall back to enexpr()
+  returning_expr <- enexpr(returning)
+  tryCatch(
+    returning_expr <- returning,
+    error = identity
+  )
+
+  returning_cols <- eval_select_both(returning_expr, colnames(x))$names
   check_returning_cols_possible(returning_cols, in_place)
 
   y <- auto_copy(x, y, copy = copy)
@@ -142,10 +166,143 @@ rows_update.tbl_dbi <- function(x, y, by = NULL, ...,
 
 #' @export
 #' @rdname rows-db
+rows_patch.tbl_dbi <- function(x, y, by = NULL, ...,
+                               in_place = NULL, copy = FALSE, check = NULL,
+                               returning = NULL) {
+
+  # Expect manual quote from user, silently fall back to enexpr()
+  returning_expr <- enexpr(returning)
+  tryCatch(
+    returning_expr <- returning,
+    error = identity
+  )
+
+  returning_cols <- eval_select_both(returning_expr, colnames(x))$names
+  check_returning_cols_possible(returning_cols, in_place)
+
+  y <- auto_copy(x, y, copy = copy)
+  y_key <- db_key(y, by)
+  by <- names(y_key)
+  x_key <- db_key(x, by)
+
+  new_columns <- setdiff(colnames(y), by)
+
+  name <- target_table_name(x, in_place)
+
+  if (!is_null(name)) {
+    # Checking optional, can rely on primary key constraint
+    if (is_true(check)) {
+      check_db_superset(x, y, by)
+    }
+
+    if (is_empty(new_columns)) {
+      return(invisible(x))
+    }
+
+    con <- dbplyr::remote_con(x)
+    sql <- sql_rows_patch(x, y, by, returning_cols = returning_cols)
+
+    rows_get_or_execute(x, con, sql, returning_cols)
+  } else {
+    # Checking optional, can rely on primary key constraint
+    # FIXME: contrary to doc currently also checks if `in_place = FALSE`
+    if (is_null(check) || is_true(check)) {
+      check_db_superset(x, y, by)
+    }
+
+    if (is_empty(new_columns)) {
+      return(x)
+    }
+
+    xy <- left_join(
+      x, y,
+      by = by,
+      suffix = c("", "...y")
+    )
+
+    patch_columns_y <- paste0(new_columns, "...y")
+    patch_quos <- lapply(new_columns, function(.x) quo(coalesce(!!sym(.x), !!sym(patch_columns_y)))) %>%
+      rlang::set_names(new_columns)
+    xy %>%
+      mutate(!!!patch_quos) %>%
+      select(-all_of(patch_columns_y))
+  }
+}
+
+#' @export
+#' @rdname rows-db
+rows_upsert.tbl_dbi <- function(x, y, by = NULL, ...,
+                                in_place = NULL, copy = FALSE, check = NULL,
+                                returning = NULL) {
+
+  # Expect manual quote from user, silently fall back to enexpr()
+  returning_expr <- enexpr(returning)
+  tryCatch(
+    returning_expr <- returning,
+    error = identity
+  )
+
+  returning_cols <- eval_select_both(enquo(returning), colnames(x))$names
+  check_returning_cols_possible(returning_cols, in_place)
+
+  y <- auto_copy(x, y, copy = copy)
+  y_key <- db_key(y, by)
+  by <- names(y_key)
+  x_key <- db_key(x, by)
+
+  new_columns <- setdiff(colnames(y), by)
+
+  name <- target_table_name(x, in_place)
+
+  if (!is_null(name)) {
+    # Checking optional, can rely on primary key constraint
+    if (is_true(check)) {
+      check_db_superset(x, y, by)
+    }
+
+    if (is_empty(new_columns)) {
+      return(invisible(x))
+    }
+
+    con <- dbplyr::remote_con(x)
+    sql <- sql_rows_upsert(x, y, by, returning_cols = returning_cols)
+
+    rows_get_or_execute(x, con, sql, returning_cols)
+  } else {
+    # Checking optional, can rely on primary key constraint
+    # FIXME: contrary to doc currently also checks if `in_place = FALSE`
+    if (is_null(check) || is_true(check)) {
+      check_db_superset(x, y, by)
+    }
+
+    existing_columns <- setdiff(colnames(x), new_columns)
+
+    unchanged <- anti_join(x, y, by = by)
+    inserted <- anti_join(y, x, by = by)
+    updated <-
+      x %>%
+      select(!!!existing_columns) %>%
+      inner_join(y, by = by)
+    upserted <- union_all(updated, inserted)
+
+    union_all(unchanged, upserted)
+  }
+}
+
+#' @export
+#' @rdname rows-db
 rows_delete.tbl_dbi <- function(x, y, by = NULL, ...,
                                 in_place = NULL, copy = FALSE, check = NULL,
                                 returning = NULL) {
-  returning_cols <- eval_select_both(enquo(returning), colnames(x))$names
+
+  # Expect manual quote from user, silently fall back to enexpr()
+  returning_expr <- enexpr(returning)
+  tryCatch(
+    returning_expr <- returning,
+    error = identity
+  )
+
+  returning_cols <- eval_select_both(returning_expr, colnames(x))$names
   check_returning_cols_possible(returning_cols, in_place)
 
   y <- auto_copy(x, y, copy = copy)
@@ -270,10 +427,35 @@ sql_rows_update <- function(x, y, by, ..., returning_cols = NULL) {
 }
 
 #' @export
+sql_rows_update.tbl_sql <- function(x, y, by, ..., returning_cols = NULL) {
+  # * avoid CTEs for the general case as they do not work everywhere
+  con <- dbplyr::remote_con(x)
+
+  p <- sql_rows_prep(x, y, by)
+
+  sql <- paste0(
+    "UPDATE ", p$name, "\n",
+    "SET\n",
+    paste0(
+      "  ", unlist(p$new_columns_qq_list),
+      " = ", unlist(p$new_columns_qual_qq_list),
+      collapse = ",\n"
+    ), "\n",
+    "FROM (\n",
+    "    ", dbplyr::sql_render(y), "\n",
+    "  ) AS ", p$y_name, "\n",
+    "WHERE (", p$compare_qual_qq, ")\n",
+    sql_returning_cols(x, returning_cols)
+  )
+
+  glue::as_glue(sql)
+}
+
+#' @export
 sql_rows_update.tbl_SQLiteConnection <- function(x, y, by, ..., returning_cols = NULL) {
   con <- dbplyr::remote_con(x)
 
-  p <- sql_rows_update_prep(x, y, by)
+  p <- sql_rows_prep(x, y, by)
 
   sql <- paste0(
     "WITH ", p$y_name, "(", p$y_columns_qq, ") AS (\n",
@@ -296,7 +478,7 @@ sql_rows_update.tbl_SQLiteConnection <- function(x, y, by, ..., returning_cols =
 `sql_rows_update.tbl_Microsoft SQL Server` <- function(x, y, by, ..., returning_cols = NULL) {
   con <- dbplyr::remote_con(x)
 
-  p <- sql_rows_update_prep(x, y, by)
+  p <- sql_rows_prep(x, y, by)
 
   # https://stackoverflow.com/a/2334741/946850
   sql <- paste0(
@@ -325,7 +507,7 @@ sql_rows_update.tbl_SQLiteConnection <- function(x, y, by, ..., returning_cols =
 sql_rows_update.tbl_MariaDBConnection <- function(x, y, by, ..., returning_cols = NULL) {
   con <- dbplyr::remote_con(x)
 
-  p <- sql_rows_update_prep(x, y, by)
+  p <- sql_rows_prep(x, y, by)
 
   # https://stackoverflow.com/a/19346375/946850
   sql <- paste0(
@@ -333,7 +515,7 @@ sql_rows_update.tbl_MariaDBConnection <- function(x, y, by, ..., returning_cols 
     "  INNER JOIN (\n", dbplyr::sql_render(y), "\n) AS ", p$y_name, "\n",
     "  ON ", p$compare_qual_qq, "\n",
     "SET\n",
-    paste0("  ", p$target_columns_qual_qq, " = ", p$new_columns_qual_qq, collapse = ",\n"),
+    paste0("  ", p$new_columns_qual_qq, " = ", p$new_columns_qual_qq, collapse = ",\n"),
     sql_returning_cols(x, returning_cols)
   )
 
@@ -344,7 +526,7 @@ sql_rows_update.tbl_MariaDBConnection <- function(x, y, by, ..., returning_cols 
 sql_rows_update.tbl_PqConnection <- function(x, y, by, ..., returning_cols = NULL) {
   con <- dbplyr::remote_con(x)
 
-  p <- sql_rows_update_prep(x, y, by)
+  p <- sql_rows_prep(x, y, by)
 
   # https://www.postgresql.org/docs/9.5/sql-update.html
   sql <- paste0(
@@ -368,10 +550,135 @@ sql_rows_update.tbl_PqConnection <- function(x, y, by, ..., returning_cols = NUL
   glue::as_glue(sql)
 }
 
-#' @export
-sql_rows_update.tbl_duckdb_connection <- sql_rows_update.tbl_SQLiteConnection
+sql_rows_upsert <- function(x, y, by, ..., returning_cols = NULL) {
+  ellipsis::check_dots_used()
+  # FIXME: check here same src for x and y? if not -> error.
+  UseMethod("sql_rows_upsert")
+}
 
-sql_rows_update_prep <- function(x, y, by) {
+# nocov start
+# exclude from coverage because no database in the workflow supports MERGE
+#' @export
+sql_rows_upsert.tbl_sql <- function(x, y, by, ..., returning_cols = NULL) {
+  con <- dbplyr::remote_con(x)
+
+  p <- sql_rows_upsert_prep(x, y, by)
+
+  update_clause <- paste0(
+    unlist(p$new_columns_qq_list), " = ", "excluded.", unlist(p$new_columns_qq_list),
+    collapse = ",\n"
+  )
+
+  sql <- paste0(
+    "MERGE INTO ", p$name, "\n",
+    "USING (", dbplyr::sql_render(y), ") AS ", p$y_name, "\n",
+    "ON (", p$compare_qual_qq, ")\n",
+    "WHEN MATCHED THEN\n",
+    "  UPDATE SET", update_clause, "\n",
+    "WHEN NOT MATCHED THEN\n",
+    "  INSERT (", p$y_columns_qq, ")\n",
+    "  VALUES (", p$y_columns_qual_qq, ")\n",
+    sql_returning_cols(x, returning_cols)
+  )
+
+  glue::as_glue(sql)
+}
+# nocov end
+
+# nocov start
+# exclude from coverage because MERGE somehow doesn't work with MS SQL 2017
+#' @export
+`sql_rows_upsert.tbl_Microsoft SQL Server` <- function(x, y, by, ..., returning_cols = NULL) {
+  con <- dbplyr::remote_con(x)
+
+  p <- sql_rows_upsert_prep(x, y, by)
+
+  update_clause <- paste0(
+    unlist(p$new_columns_qq_list), " = ", "excluded.", unlist(p$new_columns_qq_list),
+    collapse = ",\n"
+  )
+
+  sql <- paste0(
+    "MERGE INTO ", p$name, "\n",
+    "USING (", dbplyr::sql_render(y), ") AS ", p$y_name, "\n",
+    "ON (", p$compare_qual_qq, ")\n",
+    "WHEN MATCHED THEN\n",
+    "  UPDATE SET", update_clause, "\n",
+    "WHEN NOT MATCHED THEN\n",
+    "  INSERT (", p$y_columns_qq, ")\n",
+    "  VALUES (", p$y_columns_qual_qq, ")\n",
+    sql_output_cols(x, returning_cols),
+    ";"
+  )
+
+  glue::as_glue(sql)
+}
+# nocov end
+
+#' @export
+sql_rows_upsert.tbl_duckdb_connection <- function(x, y, by, ..., returning_cols = NULL) {
+  abort("upsert is not supported for DuckDB")
+}
+
+#' @export
+sql_rows_upsert.tbl_PqConnection <- function(x, y, by, ..., returning_cols = NULL) {
+  con <- dbplyr::remote_con(x)
+
+  p <- sql_rows_upsert_prep(x, y, by)
+
+  update_clause <- paste0(
+    unlist(p$new_columns_qq_list), " = ", "excluded.", unlist(p$new_columns_qq_list),
+    collapse = ",\n"
+  )
+
+  sql <- paste0(
+    "WITH ", p$y_name, "(", p$y_columns_qq, ") AS (\n",
+    dbplyr::sql_render(y),
+    "\n)\n",
+    "INSERT INTO ", p$name, " (", p$y_columns_qq, ")\n",
+    "SELECT * FROM ", p$y_name, "\n",
+    "WHERE true\n",
+    "ON CONFLICT (", p$by_columns_qq, ")\n",
+    "DO UPDATE\n",
+    "SET ", update_clause, "\n",
+    sql_returning_cols(x, returning_cols)
+  )
+
+  glue::as_glue(sql)
+}
+
+#' @export
+sql_rows_upsert.tbl_SQLiteConnection <- sql_rows_upsert.tbl_PqConnection
+
+#' @export
+sql_rows_upsert.tbl_MariaDBConnection <- function(x, y, by, ..., returning_cols = NULL) {
+  con <- dbplyr::remote_con(x)
+
+  p <- sql_rows_upsert_prep(x, y, by)
+
+  update_clause <- paste0(
+    p$new_columns_qq_list, " = ", p$new_columns_qual_qq_list,
+    collapse = ",\n"
+  )
+
+  # MariaDB has the order: first INSERT then CTE
+  # https://www.itjungle.com/2016/08/16/fhg081616-story02/
+  sql <- paste0(
+    "INSERT INTO ", p$name, "\n",
+    "WITH ", p$y_name, "(", p$y_columns_qq, ") AS (\n",
+    dbplyr::sql_render(y),
+    "\n)\n",
+    "SELECT * FROM ", p$y_name, "\n",
+    "WHERE true\n",
+    "ON DUPLICATE KEY UPDATE\n",
+    update_clause, "\n",
+    sql_returning_cols(x, returning_cols)
+  )
+
+  glue::as_glue(sql)
+}
+
+sql_rows_upsert_prep <- function(x, y, by) {
   con <- dbplyr::remote_con(x)
   name <- dbplyr::remote_name(x)
 
@@ -381,7 +688,12 @@ sql_rows_update_prep <- function(x, y, by) {
     DBI::dbQuoteIdentifier(con, colnames(y)),
     collapse = ", "
   )
+  y_columns_qual_qq <- paste(
+    y_name, ".", DBI::dbQuoteIdentifier(con, colnames(y)),
+    collapse = ", "
+  )
 
+  by_columns_qq <- DBI::dbQuoteIdentifier(con, by)
   new_columns_q <- DBI::dbQuoteIdentifier(con, setdiff(colnames(y), by))
   new_columns_qq <- paste(new_columns_q, collapse = ", ")
   new_columns_qq_list <- list(new_columns_q)
@@ -402,10 +714,125 @@ sql_rows_update_prep <- function(x, y, by) {
   tibble(
     name, y_name,
     y_columns_qq,
+    y_columns_qual_qq,
+    by_columns_qq,
     new_columns_qq, new_columns_qq_list,
     new_columns_qual_qq, new_columns_qual_qq_list,
     compare_qual_qq
   )
+}
+
+#' @export
+#' @rdname rows-db
+sql_rows_patch <- function(x, y, by, ..., returning_cols = NULL) {
+  ellipsis::check_dots_used()
+  # FIXME: check here same src for x and y? if not -> error.
+  UseMethod("sql_rows_patch")
+}
+
+#' @export
+sql_rows_patch.tbl_sql <- function(x, y, by, ..., returning_cols = NULL) {
+  # * avoid CTEs for the general case as they do not work everywhere
+  con <- dbplyr::remote_con(x)
+
+  p <- sql_rows_prep(x, y, by)
+
+  sql <- paste0(
+    "UPDATE ", p$name, "\n",
+    "SET\n",
+    paste0(
+      "  ", unlist(p$new_columns_qq_list),
+      " = ", unlist(p$new_columns_patch_qq_list),
+      collapse = ",\n"
+    ), "\n",
+    "FROM (\n",
+    "    ", dbplyr::sql_render(y), "\n",
+    "  ) AS ", p$y_name, "\n",
+    "WHERE (", p$compare_qual_qq, ")\n",
+    sql_returning_cols(x, returning_cols)
+  )
+
+  glue::as_glue(sql)
+}
+
+#' @export
+`sql_rows_patch.tbl_Microsoft SQL Server` <- function(x, y, by, ..., returning_cols = NULL) {
+  con <- dbplyr::remote_con(x)
+
+  p <- sql_rows_prep(x, y, by)
+
+  # https://stackoverflow.com/a/2334741/946850
+  sql <- paste0(
+    "WITH ", p$y_name, "(", p$y_columns_qq, ") AS (\n",
+    dbplyr::sql_render(y),
+    "\n)\n",
+    #
+    "UPDATE ", p$name, "\n",
+    "SET\n",
+    paste0(
+      "  ", unlist(p$new_columns_qq_list),
+      " = ", unlist(p$new_columns_patch_qq_list),
+      collapse = ",\n"
+    ),
+    "\n",
+    sql_output_cols(x, returning_cols),
+    "FROM ", p$name, "\n",
+    "  INNER JOIN ", p$y_name, "\n",
+    "  ON ", p$compare_qual_qq
+  )
+
+  glue::as_glue(sql)
+}
+
+#' @export
+sql_rows_patch.tbl_MariaDBConnection <- function(x, y, by, ..., returning_cols = NULL) {
+  con <- dbplyr::remote_con(x)
+
+  p <- sql_rows_prep(x, y, by)
+
+  # https://stackoverflow.com/a/19346375/946850
+  sql <- paste0(
+    "UPDATE ", p$name, "\n",
+    "  INNER JOIN (\n", dbplyr::sql_render(y), "\n) AS ", p$y_name, "\n",
+    "  ON ", p$compare_qual_qq, "\n",
+    "SET\n",
+    paste0("  ", p$new_columns_qq, " = ", p$new_columns_patch_qq, collapse = ",\n"),
+    sql_returning_cols(x, returning_cols)
+  )
+
+  glue::as_glue(sql)
+}
+
+#' @export
+sql_rows_patch.tbl_PqConnection <- function(x, y, by, ..., returning_cols = NULL) {
+  con <- dbplyr::remote_con(x)
+
+  p <- sql_rows_prep(x, y, by)
+
+  # https://www.postgresql.org/docs/9.5/sql-update.html
+  sql <- paste0(
+    "WITH ", p$y_name, " AS (\n",
+    dbplyr::sql_render(y),
+    "\n)\n",
+    #
+    "UPDATE ", p$name, "\n",
+    "SET\n",
+    paste0(
+      "  ", unlist(p$new_columns_qq_list),
+      " = ", unlist(p$new_columns_patch_qq_list),
+      collapse = ",\n"
+    ),
+    "\n",
+    "FROM ", p$y_name, "\n",
+    "WHERE ", p$compare_qual_qq,
+    sql_returning_cols(x, returning_cols)
+  )
+
+  glue::as_glue(sql)
+}
+
+sql_coalesce <- function(x, y) {
+  paste0("COALESCE(", x, ",", y, ")")
 }
 
 #' @export
@@ -416,20 +843,67 @@ sql_rows_delete <- function(x, y, by, ..., returning_cols = NULL) {
   UseMethod("sql_rows_delete")
 }
 
+sql_rows_prep <- function(x, y, by) {
+  con <- dbplyr::remote_con(x)
+  name <- dbplyr::remote_name(x)
+
+  # https://stackoverflow.com/a/47753166/946850
+  y_name <- DBI::dbQuoteIdentifier(con, "...y")
+  y_q <- DBI::dbQuoteIdentifier(con, colnames(y))
+  by_q <- DBI::dbQuoteIdentifier(con, by)
+
+  y_columns_qq <- sql_list(y_q)
+
+  new_columns_q <- setdiff(y_q, by_q)
+  new_columns_qual_q <- paste0(y_name, ".", new_columns_q)
+  old_columns_qual_q <- paste0(name, ".", new_columns_q)
+
+  new_columns_qq <- sql_list(new_columns_q)
+  new_columns_qq_list <- list(new_columns_q)
+
+  new_columns_qual_qq <- sql_list(new_columns_qual_q)
+  new_columns_qual_qq_list <- list(new_columns_qual_q)
+
+  new_columns_patch <- sql_coalesce(old_columns_qual_q, new_columns_qual_q)
+  new_columns_patch_qq <- sql_list(new_columns_patch)
+  new_columns_patch_qq_list <- list(new_columns_patch)
+
+  compare_qual_qq <- paste0(
+    y_name, ".", by_q,
+    " = ",
+    name, ".", by_q,
+    collapse = " AND "
+  )
+
+  tibble(
+    name, y_name,
+    y_columns_qq,
+    new_columns_qq, new_columns_qq_list,
+    new_columns_qual_qq, new_columns_qual_qq_list,
+    new_columns_patch_qq, new_columns_patch_qq_list,
+    compare_qual_qq
+  )
+}
+
+sql_list <- function(x) {
+  paste(x, collapse = ", ")
+}
+
 #' @export
 sql_rows_delete.tbl_sql <- function(x, y, by, ..., returning_cols = NULL) {
   con <- dbplyr::remote_con(x)
 
-  p <- sql_rows_update_prep(x, y, by)
+  p <- sql_rows_prep(x, y, by)
 
   sql <- paste0(
-    "WITH ", p$y_name, "(", p$y_columns_qq, ") AS (\n",
-    dbplyr::sql_render(y),
-    "\n)\n",
-    #
     "DELETE FROM ", p$name, "\n",
     sql_output_cols(x, returning_cols, delete = TRUE),
-    "WHERE EXISTS (SELECT * FROM ", p$y_name, " WHERE ", p$compare_qual_qq, ")",
+    "WHERE EXISTS (\n",
+    "  SELECT * FROM (\n",
+    "    ", dbplyr::sql_render(y), "\n",
+    "  ) AS ", p$y_name, "\n",
+    "  WHERE ", p$compare_qual_qq, "\n",
+    ")",
     sql_returning_cols(x, returning_cols)
   )
 
@@ -509,6 +983,11 @@ sql_returning_cols.tbl_dbi <- function(x, returning_cols, ...) {
   returning_cols <- sql_named_cols(con, returning_cols, table = dbplyr::remote_name(x))
 
   paste0("RETURNING ", returning_cols)
+}
+
+#' @export
+sql_returning_cols.tbl_duckdb_connection <- function(x, returning_cols, ...) {
+  abort("DuckDB does not support the `returning` argument.")
 }
 
 #' @export